Move scheduling helpers into own module
