+++ src/asynkit/__init__.py
@@ -8,3 +8,4 @@ from .monitor import (
     MonitorAwaitable,
     OOBData,
 )
+from .scheduling import *
